refactor: 평면 모듈·developer 도메인을 서브패키지 컨벤션에 맞춤
- progress/util/util.py → progress/util.py (중복 디렉토리 제거)
- instruction/validation/validation.py → instruction/validation.py
- developer 평면 파일 → controller/·service/ 서브패키지로 이동
  (외부 /admin/* prefix는 유지)
