+++ app/domains/progress/service/learning_record_service.py
@@ -9,7 +9,7 @@ from app.domains.progress.models import (
     WeakConcept,
 )
 from app.domains.progress.repository.learning_record_repository import LearningClassroomRepository, LearningRecordRepository
-from app.domains.progress.util.util import (
+from app.domains.progress.util import (
     _calculate_priority,
     _record_created_at,
     build_problem_key,
