+++ app/domains/progress/util.py
