+++ tests/structure/test_domain_structure.py
@@ -6,8 +6,8 @@ def test_domain_packages_are_importable():
         "app.common.security",
         "app.domains",
         "app.domains.developer",
-        "app.domains.developer.indexing_service",
-        "app.domains.developer.admin_router",
+        "app.domains.developer.service.indexing_service",
+        "app.domains.developer.controller.admin_router",
         "app.domains.auth",
         "app.domains.auth.models",
         "app.domains.auth.repository.auth_repository",
@@ -27,7 +27,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.controller.student_router",
         "app.domains.instruction.controller.teacher_router",
         "app.domains.instruction.service.generation",
-        "app.domains.instruction.validation.validation",
+        "app.domains.instruction.validation",
         "app.domains.instruction.models",
         "app.domains.instruction.repository.assignment_repository",
         "app.domains.instruction.schema.instruction_schemas",
